Fixed int/size_t problem in ProximityStrategyPQP
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@4273 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rwlibs/proximitystrategies/ProximityStrategyPQP.cpp
@@ -251,7 +251,7 @@ bool ProximityStrategyPQP::addGeometry(rw::proximity::ProximityModel* model, rw:
 bool ProximityStrategyPQP::removeGeometry(rw::proximity::ProximityModel* model, const std::string& geomId){
     PQPProximityModel *pmodel = (PQPProximityModel*) model;
     // remove from model
-    size_t idx=-1;
+    int idx=-1;
     for(size_t i=0;i<pmodel->models.size();i++)
         if(pmodel->models[i].geoid==geomId){
             idx = i;
